Don't adjust the window bounds for DPI settings
The implementation of Widget::SetBounds adjusts the given rectangle by the
screen's DPI.  This makes the WebView position not what the application expects
it to be.

There is a comment in DesktopNativeWidgetAura::SetBounds that indicates that
this scaling logic should be removed.  For now though, just expose another
method that explicitly doesn't do this scaling.  We can revisit this later to
see if this scaling logic eventually gets removed from SetBounds upstream.

(cherry picked from commit 0f44abddd663f4e7bd70a7aa7245e99dc24431aa)