Add drawContents() to WebView
This will allow the embedder to request the web view to draw on the
provided native device context. The caller needs to specify the source region
of the webview that is to be copied and the destination region of the device
context on which the final image will be drawn on. The image will scale
non-proportionally if the aspect ratio of the regions differ.

drawContents also accept a CSS class list. This class list is temporarily
applied to the body element of the document during the draw operation. This is
slightly different than CSS media selectors because it allows the caller to
specify a list of classes, instead of just a boolean state of whether the draw
operation is printing related.

The dpiMultiplier parameter allows the caller to specify an anti-alias
multiplier that the PDF engine will use when it draws raster objects into the
device context.