Always capture callstack for uncaught exceptions (for in-process renderers only)
The default behavior upstream is that uncaught exceptions will only include the
callstack when the inspector is open [1].  This commit changes that for
blpwtk2-based applications so that the callstack is captured for all uncaught
exceptions (even when the inspector is not open).  Note that this is only done
for in-process renderers.

[1] https://bugs.webkit.org/show_bug.cgi?id=42560

(cherry picked from commit e298a01fa1e20e6cce2171e3005d998b9ab6b0a1)

Conflicts:
      src/third_party/WebKit/Source/bindings/core/v8/ScriptController.h