Disable ContentMainRunner's PeekMessage hack in Toolkit-based applications
The hack in ContentMainRunner, which is used to suppress the IDC_APPSTARTING
cursor, makes the assumption that it posted the one and only message in the
queue.  This is true for chromium subprocesses, but isn't true for blpwtk2
Toolkit-based applications.

This commit disables that hack for such applications.

(cherry picked from commit c853b1fd5c6e0ac92605fb9a515195f130b1c598)