Disable resize short-circuiting when a non-standard zoom is used.
When Windows is configured with a zoom factor greater than 125%, the
dimensions of the webview need to be scaled with an inverse of the zoom
factor.  The browser process applies this inverse scale when it sends a
resize message to the renderer process.  Because we short-circuit the
event to the renderer when it is running in the main thread, this logic
was not applied and it causes the renderer to resized incorrectly.

This commit disables the short-circuiting of the resize event when a
non-standard zoom is used and rely on the round-trip of the event
through the browser process to resize the webview.