blpwtk2: Single-threaded webview
Adds the function blpwtk2::ToolkitCreateParams::enableRendererOnBrowserThread()
which, when the toolkit's thread mode is set to 'ThreadMode::ORIGINAL', will
cause Chromium's renderer to run on the same thread as the browser.  The
purpose of this is to facilitate implementing HWND behavior with JavaScript
that can access renderer state (such as the DOM).