Check the custom dictionary when we encounter a misspelling
This change was lost when merging the upstream implementation of the
multilingual spellchecker.  When we encounter a misspelling, we should check
the custom dictionary.  If the word is found in the custom dictionary, then let
it continue searching for other misspelled words.

(cherry picked from commit 6720435aa6bece3939fb6e578bcd3f8873efda3d)