Make single-threaded blpwtk2 into a thread mode
Prior to this commit, the embedder enabled single-threaded mode in
blpwtk2 by setting the thread mode to ORIGINAL (meaning that the browser
thread runs on the application thread) and by setting the
enableRendererOnBrowserThread flag.  This commit simplifies the process
by introducing a third thread mode SINGLE.

This commit also contains some code cleanup and refactor.  More common
code is moved from WebViewProxy to RendererUtil so WebViewImpl needn't
have a separate copy of it. Dependency on DWM and some unused functions
are also removed.