Break spinlock on internal window message.
This commit prevents the spinlock of the message loop when an internal
message is added into the windows message queue.

GetQueueStatus() may return a flag indicating the presence of input
messages in the message queue but it may not be a message that can be
fetched with GetMessage() or PeekMessage().  The intended use of
GetQueueStatus() is to do a fast check of the queue.  If it returns
true, we need to check further with a more expensive call to
PeekMessage() in no-remove mode.