Introduce the interface 'blpwtk2::WebViewHostObserver'
The purpose of this interface is to allow an application running in the browser thread to keep track of the browser-side implementations of WebViews. The functions 'webViewHostCreated()' and 'webViewHostDestroyed()' will be called when the webview are created and destroyed. The browser-side objects are identified by a string (the 'channelId' used by the renderer process to communicate with the browser process) and by an integer (the 'routingId' that is generated when the renderer-side of a webview is created).