Provide the URL loader with the HTTP request body
This change adds a new 'request()' function to the RequestContext class.
This accessor function returns a pointer to a URLRequest object which
contain details of the HTTP request, including the full body message of
the HTTP request. This will enable the embedder's URL loader to get
these HTTP request details.