Rubberband: Conservative expansion of selection area
Prior to this commit, the rubberband area expanded downward when there
were overlapping web elements.  This made text selection within a single
element impossible if another element was slightly overlapping it.  This
commit addresses the problem by checking if the selection candidate's
bottom edge is within the bounds of the mouse selection box.