Always disable spellchecker auto-correction
The preference dictionary does not have a setting for auto-correction
and yet the spellcheck service attempts to check the value of the
auto-correction flag.  This will fail in debug build because of an
assertion trip but the value will default to false in release builds.
This commit hard-codes the value to false to prevent the assertion trip.