InProcessURLRequest: Allow for long-lived references to data members
Prior to this commit, the references passed to the class were held onto
with raw pointers.  This doesn't work well because the lift time for these
references are relatively short lived and they need to survive async
callbacks.  This change extracts the information from content::RequestInfo
for local class storage and keeps a shared pointer reference to
content::ResourceRequestBody.