Inspector: Make sure all buffer fragments are processed
Prior to this commit, the browser thread in the inspector's render process
shipped only the first response fragment to the UI thread.  This is fixed
so that it now waits for the entire message to arrive and then ships it off
to the UI thread, which then gets passed into the inspector's JS code.