BrokerAttachmentPrivileged: Add thread-safety for untrusted senders in browser
When an IPC call is issued, any attached handle in it are mangled to ensure
that the two peers are unaware of each other's handle ids.  This mangling is
done by the privileged attachment broker living in the browser process.  The
untrusted processes only have access to their unprivileged attachment broker,
which merely proxies the request to the browser's privileged broker.

Since the broker is registered process-wide, a renderer that runs inside of
the browser process will be able to use the privileged broker directly.  This
would be fine if the broker is thread-safe but unfortunately it's not.  Since
the privileged broker is owned by the IO thread of the browser
(aka. Chrome_IOThread), any attempt to use it from the IO thread of the
renderer (aka. Chrome_ChildIOThread) will likely corrupt the thread-unsafe IPC
channel.

This commit adds some thread-safety to the privileged broker so that a 'send'
operation will be dispatched from Chrome_ChildIOThread to Chrome_IOThread
before touching the IPC channel.  The Chrome_ChildIOThread will wait until the
operation is complete since the operation is required to be synchronous.