Clip the rectangle used to bound the 'ghost image' generated during mouse drag
The function blink::LayoutObject::absoluteBoundingBoxRectIncludingDescendants()
is used to compute the rectangle used to bound content when generating a drag
image. Prior to this change, this function would ignore any 'overflow' properties
set on the object, which would generate incorrect (usually too-large)
bounding boxes for the generated image.

This issue has been identified upstream, but not fixed:
https://bugs.chromium.org/p/chromium/issues/detail?id=634348
https://bugs.chromium.org/p/chromium/issues/detail?id=482383