Avoid blurry content due to composition layer transforms
Chrome's compositor will decide to apply bilinear image filtering to the layers
that it composites if those layers have subpixel translation or non-uniform
scaling transformations applied to them. This can make the content of those
layers, especially text, appear blurry. These changes seek to avoid that by
applying such transformations when the layers are rasterized rather than when
they are composited.

The upstream compositor already extracts the uniform scaling term, represented
by a single 'float' value, and applies it at rasterization time. These changes
replace that value with an instance of a new type, 'gfx::AxisTransform2d',
which represents translation and non-uniform scaling. This transform is
extracted from the cc::PictureLayer's 'to-screen' transformation and then
passed down the stack of function calls that ultimately rasterize the layer's
display list (using the Skia library). This transform is also removed from
the transform added to each quad that is rendered by the compositor's
cc::GLRenderer component; since the quads will not have subpixel or
non-uniform scaling, cc::GLRenderer will not use bilinear filtering when
reading the layer's texture.

Some of these changes are equivalent to work taking place upstream to
address issue https://bugs.chromium.org/p/chromium/issues/detail?id=521364.
That work introduces the 'gfx::AxisTransform2d' type that is passed down to
the rasterizer in place of the single 'float' type.

The remainder of these changes extend the upstream work so that non-uniform
scaling is applied at rasterization time in addition to just uniform scaling.
Those changes do not have an upstream equivalent, and since the compositor has
mostly been biased toward representing a layer's uniform scaling only (it
extracted the larger of the two scaling terms and used that as the single
scaling term passed to the rasterizer), another new type, 'gfx::Uniform2d',
is used to represent non-uniform scaling throughout the compositor (mostly
in the components that implement a cc::PictureLayer's tiling).