Ignore characteristics of the layer's transform matrix when disabling LCD text
The main purpose of this change is so that LCD text will be enabled in the case
where a layer has a scaling transform applied to it. This change will actually
permit LCD text with rotational and perspective transforms, which are still
undesirable, so will be revisited as part of the raster scaling work.