Don't double close the shared memory handle
The commit ad6f6d8a changes DuplicateHandle() so that it passes the ownership
of the handle to the IPC channel.  This allows the programmer to no longer
worry about closing the handle after the IPC communication is complete.

In this case, the one call to close the handle was accidentally not removed
so the handle was closed twice; Once by the IPC channel object and another by
PrintWebViewHelper::PrintPagesNative. This commit removes the call to close
the handle from PrintPagesNative. This change was also introduced in future
versions of Chromium:
 commit 741ae823 (from Chromium 53)
 commit ae78ab89 (from Chromium 54)

This commit also removes the code that enables message pumping while
initializing the print settings.  The enablement of the message pump
was only needed in single-threaded mode to avoid a deadlock. An alternative
solution will be applied later that works for single-threaded mode.