Conditionally pump messages when fetching default print settings
This commit is reapplying a previously applied patch that was reverted.
When the renderer sends a sync IPC message to the browser, the renderer
blocks until the browser responds. This is fine in multi-process mode
but it introduces a deadlock in single-threaded mode. To work around the
deadlock, the IPC channel is flagged to allow pumping of messages while
the renderer is waiting for the browser.  This fixes the issue for
single-threaded mode but introduces a crash in multi-process mode starting
from Chromium 50. This commit is enabling the message pumping during the
wait for the IPC reply only when the toolkit is operating in single-threaded
mode.